refactor: folder page index.page.tsx 임시방편으로 layout 폴더로 로직 분리
