+++ pages/folder/comp/article/comp/card-container/comp/folder-link-category/hooks/useGetFolderCategoryList.ts
@@ -1,10 +1,13 @@
-// TODO: 페이지 로직 정리되면 파일 지우기
 import { useState } from 'react';
 
 import { FolderCategoryDataWithIdTotal } from '@pages/folder/index.page';
 
 import { getFolderCategory } from '@api/folder-page/getFolderCategory';
 
+/**
+ *
+ * @todo 이거 state랑 fetch 코드 전부 FolderPageServerSidePropContextProvider로 옮겨도 될 듯.
+ */
 const useGetFolderCategoryList = (initialFolderCategoryList: FolderCategoryDataWithIdTotal[]) => {
   const [folderCategoryList, setFolderCategoryList] =
     useState<FolderCategoryDataWithIdTotal[]>(initialFolderCategoryList);
