+++ pages/folder/index.page.tsx
@@ -1,16 +1,16 @@
-import { useMemo, useRef } from 'react';
+import { ParsedUrlQuery } from 'querystring';
 
 import { GetServerSideProps, InferGetServerSidePropsType } from 'next';
 import Head from 'next/head';
 
 import { getFolderCategory, TFolderCategoryData } from '@api/folder-page/getFolderCategory';
-import { useTargetVisibleState } from '@hooks/useTargetVisibleState';
 
-import AddLink from './comp/add-link/AddLink';
-import Article from './comp/article/Article';
-import Footer from './comp/footer/Footer';
-import Header from './comp/header/Header';
 import FolderPageServerSidePropContextProvider from './context/folderPageContext';
+import Layout from './layout/Layout';
+
+export interface FolderPageUrlQuery extends ParsedUrlQuery {
+  userId: string;
+}
 
 export interface FolderCategoryDataWithIdTotal extends Partial<Omit<TFolderCategoryData, 'id' | 'name'>> {
   id: number | 'total';
@@ -24,12 +24,12 @@ interface ServerSideProps {
 // 실제로는 로그인한 유저의 userId를 받아서 그 유저가 저장한 폴더 카테고리를 보여주도록 해야 함.
 export const getServerSideProps: GetServerSideProps<ServerSideProps> = async (context) => {
   // TODO: userId는 로그인한 유저의 id를 받아와야 함. 나중에 default 1도 없애셈.
-  const { userId = '1' } = context.query;
+  const { userId = '1' } = context.query as FolderPageUrlQuery;
 
   let folderCategoryList: FolderCategoryDataWithIdTotal[] = [];
 
   try {
-    const res = await getFolderCategory(userId as string);
+    const res = await getFolderCategory(userId);
 
     if (res) {
       const { data } = res;
@@ -51,37 +51,14 @@ export const getServerSideProps: GetServerSideProps<ServerSideProps> = async (co
   };
 };
 
-const FolderPage = <T extends HTMLElement>({
-  folderCategoryList,
-}: InferGetServerSidePropsType<typeof getServerSideProps>) => {
-  const addLinkRef = useRef<T | null>(null);
-  const footerRef = useRef<T | null>(null);
-
-  const isAddLinkVisible = useTargetVisibleState(addLinkRef, ({ target }) => {
-    if (target && target.clientHeight <= document.documentElement.scrollTop) {
-      return false;
-    }
-
-    return true;
-  });
-
-  const isFooterVisible = useTargetVisibleState(footerRef);
-
-  const shouldAddLinkLocateBottom = useMemo(
-    () => !isAddLinkVisible && !isFooterVisible,
-    [isAddLinkVisible, isFooterVisible],
-  );
-
+const FolderPage = ({ folderCategoryList }: InferGetServerSidePropsType<typeof getServerSideProps>) => {
   return (
     <>
       <Head>
         <title>Linkbrary folder page</title>
       </Head>
       <FolderPageServerSidePropContextProvider folderCategoryList={folderCategoryList}>
-        <Header />
-        <AddLink ref={addLinkRef} shouldAddLinkLocateBottom={shouldAddLinkLocateBottom} />
-        <Article />
-        <Footer footerRef={footerRef} />
+        <Layout />
       </FolderPageServerSidePropContextProvider>
     </>
   );
